Moved identity source validator into validators folder
