+++ samcli/commands/local/lib/validators/identity_source_validator.py
